+++ codegen/src/template/deno/gen/[[table]]/[[table]].service.ts
@@ -272,7 +272,23 @@ async function updateById(
  */
 async function deleteByIds(
   ids: string[],
-): Promise<number> {
+): Promise<number> {<#
+  if (hasLocked) {
+  #>
+  
+  const lockedIds: string[] = [ ];
+  for (let i = 0; i < ids.length; i++) {
+    const id = ids[i];
+    const is_locked = await <#=table#>Dao.getIs_lockedById(id);
+    if (is_locked) {
+      lockedIds.push(id);
+    }
+  }
+  if (lockedIds.length > 0 && lockedIds.length === ids.length) {
+    throw await ns("不能删除已经锁定的数据");
+  }<#
+  }
+  #>
   const data = await <#=table#>Dao.deleteByIds(ids);<#
   if (table === "i18n") {
   #>
